Stage 1: Fix schema drift between SQLAlchemy models and Alembic migrations (#22)
* Initial plan

* Stage 1 schema alignment: add missing tables/columns, fix FK types, add tests


* Fix index name: ix_product_id → ix_payments_product_id for consistency


* Use pgvector Postgres image to support vector extension in migrations


---------

+++ backend/app/db/models/rag_chunk.py
@@ -35,7 +35,7 @@ class RagChunk(Base):
     )
 
     rag_item_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("rag_items.id", ondelete="CASCADE"),
         nullable=False,
         index=True,
